LayoutTests/editing/deleting/delete-br-012-expected.txt

This tests pressing delete key to remove empty quoted line places cursor at the right location. rdar://problem/4813973
| "\n"
| <div>
|   class="editing"
|   "\n    "
|   <div>
|     style="border-style:solid; border-color:black;"
|     "\n        hello\n    "
|   "\n    "
| <div>
|   id="test"
|   style="border-style:solid; border-color:blue;"
|   <#selection-caret>
|   <br>
|   "\n"
| "\n\n\n"
